requirements.txt
streamlit
langchain
langchain-core
langchain-groq
supabase
python-dotenv
fpdf2
requests
